Initial reimplementation of DrawManager for draw logic
